LayoutTests/inspector/debugger/setPauseOnAssertions-expected.txt

CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
Debugger.setPauseOnAssertions


== Running test suite: Debugger.setPauseOnAssertions
-- Running test case: Debugger.setPauseOnAssertions.Disabled
PASS: Should not pause.

-- Running test case: Debugger.setPauseOnAssertions.Enabled
PASS: Should paused.

-- Running test case: Debugger.setPauseOnAssertions.Options.Condition

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: Debugger.setPauseOnAssertions.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: Debugger.setPauseOnAssertions.Options.IgnoreCount

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: Debugger.setPauseOnAssertions.Options.Action.Log

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: Debugger.setPauseOnAssertions.Options.Actions.Evaluate

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: Debugger.setPauseOnAssertions.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
